+++ fixtures/ios_app_with_static_frameworks/Modules/C/Sources/C.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public class C {
+    public static let value: String = "cValue"
+
+    public static func printFromC() {
+        print("print from C")
+    }
+}