+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 - Supporting bundle target dependencies that reside in different projects (in `TuistGenerator`) https://github.com/tuist/tuist/pull/348 by @kwridan 
 - Fixing header paths including folders and non-header files https://github.com/tuist/tuist/pull/356 by @kwridan
 - Fix duplicate localized resource files https://github.com/tuist/tuist/pull/363 by @kwridan
+- Update static dependency lint rule https://github.com/tuist/tuist/pull/360 by @kwridan
 
 ## 0.14.0